Use rtype to clarify types returned because of __get__()
`UnaryStruct` uses __get__() to return a different type, but this is unclear currently because now that these variable have type annotations, class variable utilizing ot are automatically typed as `UnaryStruct`.

While one option is to manually document the class variables, I think it's preferable to use an attached docstring so the documentation and code don't ever get split up if one is removed for some reason.  Using Sphinx `rtype` seems to be the best way to document it then.